+++ test/components/hierarchy/hierarchy.e2e-spec.js
@@ -39,6 +39,8 @@ describe('Hierarchy index tests', () => {
 describe('Hierarchy context menu tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/hierarchy/example-context-menu-with-details?layout=nofrills');
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.id('hierarchy'))), config.waitsFor);
     await browser.driver.sleep(config.sleep);
   });